backend/requirements.txt
flask
flask-cors
face_recognition
opencv-python
python-dotenv
